pyproject.toml
[tool.poetry]
name = "sprite-unpack"
version = "0.1.0"
description = "Tool for cutting spritesheets into seperate images for importing into Godot, Unity, etc."
authors = ["Ilya Averyanov <av@rubybox.ru>"]
keywords = []
readme = "README.md"
license = "MIT"
homepage = "https://github.com/savonarola/sprite_unpack"
repository = "https://github.com/savonarola/sprite_unpack"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"
tqdm = "^4.60.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"
mypy = "^0.812"
flake8 = "^3.9.1"
black = {version = "^20.8b1", allow-prereleases = true}

[tool.poetry.scripts]
sprite-unpack = 'sprite_unpack.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
